# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Segments API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/log/{userID}": {
            "get": {
                "description": "Получение истории добавления и удаления сегментов указанного пользователя за определенный год и месяц в формате CSV.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Получить историю изменения сегментов пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "year-month",
                        "description": "filter date",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        },
        "/segment": {
            "post": {
                "description": "Метод создания сегмента. Принимает slug (название) сегмента. Опционально можно указать процент пользователей, которые добавятся в этот сегмент автоматически.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Создать новый сегмент",
                "parameters": [
                    {
                        "description": "segment name and user percentage (optional)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_segment.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "(optional) segment name and added users",
                        "schema": {
                            "$ref": "#/definitions/create_segment.response"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        },
        "/segment/{slug}": {
            "delete": {
                "description": "Метод удаления сегмента. Принимает slug (название) сегмента.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Удалить сегмент",
                "parameters": [
                    {
                        "type": "string",
                        "description": "segment name",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Метод создания пользователя. Принимает на вход id пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Создать нового пользователя",
                "parameters": [
                    {
                        "description": "user id",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_user.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        },
        "/user-segments": {
            "post": {
                "description": "Метод изменения активных сегментов пользователя. Принимает список slug (названий) сегментов которые нужно добавить пользователю, список slug (названий) сегментов которые нужно удалить у пользователя, id пользователя. Также есть возможность задать TTL для добавляемых сегментов, чтобы по истечению времени они автоматически удалились у пользователя. TTL задается в формате \"1y8m21d\". Если хотите только удалить определенные сегменты, то можно опустить список для добавления и наоборот.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Изменить сегменты пользователя",
                "parameters": [
                    {
                        "description": "user id, segment's list to add (with ttl optional), segment's list to delete",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/change_user_segments.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of changes",
                        "schema": {
                            "$ref": "#/definitions/change_user_segments.response"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        },
        "/user-segments/{userID}": {
            "get": {
                "description": "Метод получения активных сегментов пользователя. Принимает на вход id пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получить активные сегменты пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of segments",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        },
        "/user/{userID}": {
            "delete": {
                "description": "Метод удаления пользователя. Принимает на вход id пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Удалить пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    },
                    "default": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/handlers.responseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "change_user_segments.request": {
            "type": "object",
            "properties": {
                "to_add": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/change_user_segments.segmentWithTTL"
                    }
                },
                "to_delete": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "change_user_segments.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "operation_type": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "change_user_segments.segmentWithTTL": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                },
                "ttl": {
                    "type": "string"
                }
            }
        },
        "create_segment.request": {
            "type": "object",
            "properties": {
                "percentage": {
                    "type": "number"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "create_segment.response": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                },
                "users_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "create_user.request": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.responseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}